chore(suite): move coinjoin related types from @suite-common/wallet-types
