+++ packages/suite/src/actions/wallet/coinjoinClientActions.ts
@@ -11,7 +11,8 @@ import * as COINJOIN from './constants/coinjoinConstants';
 import { breakdownCoinjoinBalance, prepareCoinjoinTransaction } from '@wallet-utils/coinjoinUtils';
 import { CoinjoinClientService } from '@suite/services/coinjoin/coinjoinClient';
 import { Dispatch, GetState } from '@suite-types';
-import { Account, CoinjoinServerEnvironment, RoundPhase } from '@suite-common/wallet-types';
+import { Account } from '@suite-common/wallet-types';
+import { CoinjoinServerEnvironment, RoundPhase } from '@wallet-types/coinjoin';
 import { onCancel as closeModal, openModal } from '@suite-actions/modalActions';
 import { notificationsActions } from '@suite-common/toast-notifications';
 import { selectAccountByKey } from '@suite-common/wallet-core';
