+++ packages/suite/src/components/suite/CoinjoinStatusBar.tsx
@@ -3,7 +3,8 @@ import styled from 'styled-components';
 import { useDispatch } from 'react-redux';
 import { Button, variables } from '@trezor/components';
 import { selectAccountByKey } from '@suite-common/wallet-core';
-import { CoinjoinSession, WalletParams } from '@suite-common/wallet-types';
+import { WalletParams } from '@suite-common/wallet-types';
+import { CoinjoinSession } from '@wallet-types/coinjoin';
 import { COINJOIN_PHASE_MESSAGES } from '@suite-constants/coinjoin';
 import { selectDevice } from '@suite-actions/suiteActions';
 import { goto } from '@suite-actions/routerActions';
