+++ packages/suite/src/types/wallet/coinjoin.ts
